Deletes legacy resources code
