+++ harvester/core/fixtures/resources-basic-delta.json
@@ -1,50 +1,5 @@
 [
     {
-        "model": "core.fileresource",
-        "pk": 16942,
-        "fields": {
-            "uri": "surfsharekit.nl/objectstore/48fac5c4-c846-48b3-9318-6b01e151a508",
-            "status": 200,
-            "config": "{\"_namespace\": \"global\", \"_private\": [\"_private\", \"_defaults\", \"_namespace\"], \"resource\": \"core.FileResource\", \"backoff_delays\": [2, 4, 8, 16]}",
-            "created_at": "2021-06-24T13:00:54.569Z",
-            "modified_at": "2021-06-24T13:00:54.569Z",
-            "purge_at": "2021-07-23T10:34:07.514Z",
-            "retainer_type": null,
-            "retainer_id": null,
-            "data_hash": "",
-            "request": {
-                "url": "https://surfsharekit.nl/objectstore/48fac5c4-c846-48b3-9318-6b01e151a508",
-                "args": [
-                    "https://surfsharekit.nl/objectstore/48fac5c4-c846-48b3-9318-6b01e151a508"
-                ],
-                "data": null,
-                "cancel": false,
-                "kwargs": {},
-                "method": "get",
-                "headers": {
-                    "Accept": "*/*",
-                    "Connection": "keep-alive",
-                    "User-Agent": "DataGrowth (v0.17.1); python-requests/2.24.0",
-                    "Accept-Encoding": "gzip, deflate"
-                },
-                "backoff_delay": false
-            },
-            "head": {
-                "Date": "Wed, 23 Jun 2021 10:34:07 GMT",
-                "Server": "Apache/2.4.37 (centos) OpenSSL/1.1.1g",
-                "Connection": "Keep-Alive",
-                "Keep-Alive": "timeout=5, max=83",
-                "Content-Type": "application/pdf",
-                "X-Powered-By": "PHP/7.2.24",
-                "Accept-Ranges": "none",
-                "Last-Modified": "Thu, 01 Jan 1970 00:00:00 GMT",
-                "Content-Length": "7551293",
-                "Content-Disposition": "attachment; filename=file/48fac5c4-c846-48b3-9318-6b01e151a508/Week-3_9-changes-in-the-labour-market.pdf"
-            },
-            "body": "test/20210623103408045217.48fac5c4-c846-48b3-9318-6b01e151a508.html"
-        }
-    },
-    {
         "model": "core.httptikaresource",
         "pk": 50,
         "fields": {
