+++ harvester/core/tasks/__init__.py
@@ -1,5 +1,2 @@
-from .generate_browser_preview import generate_browser_preview
-from .generate_pdf_preview import generate_pdf_preview
-from .generate_youtube_preview import generate_youtube_preview
 from .commands import clean_data
 from .elastic_search import sync_indices
