+++ harvester/core/tests/factories.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from django.conf import settings
 
 from core.models import (Document, Collection, Dataset, DatasetVersion, Harvest, HarvestSource, ElasticIndex,
-                         FileResource, HttpTikaResource)
+                         HttpTikaResource)
 from core.constants import HarvestStages, Repositories, DeletePolicies
 
 
@@ -110,13 +110,6 @@ class HarvestFactory(factory.django.DjangoModelFactory):
     dataset = factory.SubFactory(DatasetFactory)
     source = factory.SubFactory(HarvestSourceFactory)
     stage = HarvestStages.NEW
-
-
-class FileResourceFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = FileResource
-
-    uri = "https://maken.wikiwijs.nl/124977/Zorgwekkend_gedrag___kopie_1"
 
 
 class HttpTikaResourceFactory(factory.django.DjangoModelFactory):
